tools/crypto-balance/requirements.txt
# 자동매매 / 잔고 조회 / 웹 대시보드
ccxt>=4.0.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
google-genai>=1.0.0
